+++ client/src/app/file-browser/services/filesystem.service.ts
@@ -359,8 +359,8 @@ export class FilesystemService {
 
   acquireLock(hashId: string, options: { own: true }): Observable<ObjectLock[]> {
     return this.http.put<ResultList<ObjectLockData>>(
-      `/api/filesystem/objects/${encodeURIComponent(hashId)}/locks?own=true`,
-      {},
+      `/api/filesystem/objects/${encodeURIComponent(hashId)}/locks`,
+      { own: true },
       this.apiService.getHttpOptions(true),
     ).pipe(
       map(data => {
@@ -381,8 +381,9 @@ export class FilesystemService {
 
   deleteLock(hashId: string, options: { own: true }): Observable<{}> {
     return this.http.delete<unknown>(
-      `/api/filesystem/objects/${encodeURIComponent(hashId)}/locks?own=true`, {
+      `/api/filesystem/objects/${encodeURIComponent(hashId)}/locks`, {
         ...this.apiService.getHttpOptions(true),
+        body: { own: true },
       },
     ).pipe(
       map(() => ({})),